+++ Commands/Provisioning/SaveProvisioningTemplate.cs
@@ -14,11 +14,11 @@ using System.Threading.Tasks;
 namespace SharePointPnP.PowerShell.Commands.Provisioning
 {
     [Cmdlet(VerbsData.Save, "PnPProvisioningTemplate")]
-    [CmdletHelp("Saves a PnP file to the file systems",
+    [CmdletHelp("Saves a PnP provisioning tempalte to the file system",
         Category = CmdletHelpCategory.Provisioning)]
     [CmdletExample(
        Code = @"PS:> Save-PnPProvisioningTemplate -InputInstance $template -Out .\template.pnp",
-       Remarks = "Saves a PnP file to the file systems",
+       Remarks = "Saves a PnP provisioning template to the file system as a PnP file.",
        SortOrder = 1)]
     public class SaveProvisioningTemplate : PSCmdlet
     {